requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
Authlib==1.5.2
bcrypt==4.2.1
black==25.1.0
branca==0.8.1
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
cryptography==44.0.0
distlib==0.3.9
dnspython==2.7.0
dulwich==0.22.7
email_validator==2.2.0
fastapi==0.115.6
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
fastjsonschema==2.21.1
filelock==3.16.1
folium==0.19.5
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
gpxpy==1.6.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-oauth==0.16.1
idna==3.10
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.5
keyring==25.6.0
makefun==1.15.6
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
mypy==1.14.1
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
phonenumbers==9.0.1
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.0
poetry-core==2.0.0
pwdlib==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
pyright==1.1.391
pytest==8.3.4
pytest-asyncio==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.11.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.1
watchfiles==1.0.4
websockets==15.0.1
WTForms==3.1.2
xyzservices==2025.1.0
